+++ src/main/kotlin/cn/yiiguxing/plugin/translate/ui/DeeplConfigurationDialog.kt
@@ -96,6 +96,7 @@ class DeeplConfigurationDialog : DialogWrapper(false) {
         }
     }
 
+    @Suppress("DuplicatedCode")
     private fun createHintPane(): JComponent = JEditorPane().apply {
         isEditable = false
         isFocusable = false